Refactored files
