+++ ui/flts/workflow_manager/workflow_manager_widget.py
@@ -19,11 +19,11 @@ copyright            : (C) 2019
 """
 from PyQt4.QtGui import *
 from sqlalchemy import exc
-from sqlalchemy.orm import joinedload
-from stdm.ui.flts.workflow_manager.config import SchemeConfig
 from stdm.ui.flts.workflow_manager.config import StyleSheet
-from stdm.data.configuration import entity_model
 from stdm.settings import current_profile
+from stdm.ui.flts.workflow_manager.data_service import (
+    SchemeDataService,
+)
 from stdm.ui.flts.workflow_manager.model import WorkflowManagerModel
 from stdm.ui.flts.workflow_manager.ui_workflow_manager import Ui_WorkflowManagerWidget
 
@@ -139,62 +139,3 @@ class WorkflowManagerWidget(QWidget, Ui_WorkflowManagerWidget):
         else:
             widgets.setEnabled(False)
 
-
-class SchemeDataService:
-    """
-    Scheme data model services
-    """
-    def __init__(self, current_profile):
-        self._profile = current_profile
-        self.entity_name = "Scheme"
-
-    @property
-    def field_option(self):
-        """
-        Scheme field option
-        :return: Column and query field options
-        :rtype: List
-        """
-        return SchemeConfig().field_option
-
-    def related_entity_name(self):
-        """
-        Related entity name
-        :return entity_name: Related entity name
-        :rtype entity_name: List
-        """
-        entity_name = []
-        for fk in self._entity_model().__table__.foreign_keys:
-            entity_name.append(fk.column.table.name)
-        return entity_name
-
-    def run_query(self):
-        """
-        Run query on an entity
-        :return query_object: Query results
-        :rtype query_object: Query
-        """
-        model = self._entity_model()
-        entity_object = model()
-        try:
-            query_object = entity_object.queryObject(). \
-                options(joinedload(model.cb_check_lht_land_rights_office)). \
-                options(joinedload(model.cb_check_lht_region)). \
-                options(joinedload(model.cb_check_lht_relevant_authority)). \
-                options(joinedload(model.cb_cdrs_title_deed)).order_by(model.date_of_approval)
-            return query_object
-        except (exc.SQLAlchemyError, Exception) as e:
-            raise e
-
-    def _entity_model(self):
-        """
-`       Scheme entity model
-        :return model: Scheme entity model;
-        :rtype model: DeclarativeMeta
-        """
-        try:
-            entity = self._profile.entity(self.entity_name)
-            model = entity_model(entity)
-            return model
-        except AttributeError as e:
-            raise e
